+++ packages/eslint-plugin/tests/rules/prefer-optional-chain.test.ts
@@ -276,7 +276,7 @@ ruleTester.run('prefer-optional-chain', rule, {
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`foo?.bar?.baz || baz && baz.bar && baz.bar.foo`,
+              output: `foo?.bar?.baz || baz && baz.bar && baz.bar.foo`,
             },
           ],
         },
@@ -285,7 +285,7 @@ ruleTester.run('prefer-optional-chain', rule, {
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`foo && foo.bar && foo.bar.baz || baz?.bar?.foo`,
+              output: `foo && foo.bar && foo.bar.baz || baz?.bar?.foo`,
             },
           ],
         },
@@ -367,7 +367,7 @@ ruleTester.run('prefer-optional-chain', rule, {
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`foo?.["some long string"]?.baz`,
+              output: `foo?.["some long string"]?.baz`,
             },
           ],
         },
@@ -382,7 +382,7 @@ ruleTester.run('prefer-optional-chain', rule, {
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`foo?.[\`some long string\`]?.baz`,
+              output: `foo?.[\`some long string\`]?.baz`,
             },
           ],
         },
@@ -409,7 +409,7 @@ ruleTester.run('prefer-optional-chain', rule, {
 foo && foo.bar(/* comment */a,
   // comment2
   b, );
-      `.trimRight(),
+      `,
       output: null,
       errors: [
         {
@@ -417,11 +417,11 @@ foo && foo.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`
+              output: `
 foo?.bar(/* comment */a,
   // comment2
   b, );
-              `.trimRight(),
+      `,
             },
           ],
         },
@@ -755,7 +755,7 @@ foo?.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`if (foo) { foo?.bar; }`,
+              output: `if (foo) { foo?.bar; }`,
             },
           ],
         },
@@ -771,7 +771,7 @@ foo?.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`if (foo?.bar) { foo.bar; }`,
+              output: `if (foo?.bar) { foo.bar; }`,
             },
           ],
         },
@@ -988,7 +988,7 @@ foo?.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`if (foo) { foo?.bar; }`,
+              output: `if (foo) { foo?.bar; }`,
             },
           ],
         },
@@ -1004,7 +1004,7 @@ foo?.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`if (foo?.bar) { foo.bar; }`,
+              output: `if (foo?.bar) { foo.bar; }`,
             },
           ],
         },
@@ -1052,7 +1052,7 @@ foo?.bar(/* comment */a,
           suggestions: [
             {
               messageId: 'optionalChainSuggest',
-              output: noFormat`((typeof x) as string)?.bar;`,
+              output: `((typeof x) as string)?.bar;`,
             },
           ],
         },